% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cima-.r
\name{cima}
\alias{cima}
\title{Calculating Confidence Intervals}
\usage{
cima(y, se, alpha = 0.05, method = c("boot", "DL", "HK", "SJ"),
  B = 25000, maxit1 = 1e+05, eps = 10^(-10), lower = 0,
  upper = 1000, maxit2 = 1000, tol = .Machine$double.eps^0.25,
  rnd = NULL, maxiter = 100)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}

\item{alpha}{the alpha level of the prediction interval}

\item{method}{the calculation method for the pretiction interval (default = "boot").
\itemize{
\item \code{boot}: A parametric bootstrap confidenceS interval
           (Nagashima et al., 2018).
\item \code{DL}: A Wald-type confidence interval
           (the DerSimonian & Laird estimator for \eqn{\tau^2} with
           a standard SE estimator for the average effect,
           \eqn{(1/\sum{\hat{w}_i})^{-1}}).
\item \code{HK}: A Wald-type t-distribution confidence interval
           (the REML estimator for \eqn{\tau^2} with
           the Hartung and Knapp (2001)'s SE estimator for the average effect).
\item \code{SJ}: A Wald-type t-distribution confidence interval
           (the REML estimator for \eqn{\tau^2} with
           the Sidik and Jonkman (2006)'s bias coreccted SE estimator
           for the average effect).
}}

\item{B}{the number of bootstrap samples}

\item{maxit1}{the maximum number of iteration for the exact distribution function of \eqn{Q}}

\item{eps}{the desired level of accuracy for the exact distribution function of \eqn{Q}}

\item{lower}{the lower limit of random numbers of \eqn{\tau^2}}

\item{upper}{the lower upper of random numbers of \eqn{\tau^2}}

\item{maxit2}{the maximum number of iteration for numerical inversions}

\item{tol}{the desired level of accuracy for numerical inversions}

\item{rnd}{a vector of random numbers from the exact distribution of \eqn{\tau^2}}

\item{maxiter}{the maximum number of iteration for REML estimation}
}
\value{
\itemize{
\item \code{muhat}: the average treatment effect estimate \eqn{\hat{\mu}}.
\item \code{lci}, \code{uci}: the lower and upper confidence limits \eqn{\hat{\mu}_l} and \eqn{\hat{\mu}_u}.
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
This function calculates confidence intervals.
}
\examples{
data(sbp, package = "pimeta")
set.seed(20161102)
\donttest{pimeta::pima(sbp$y, sbp$sigmak, B = 25000)}
}
\references{
Nagashima, K., Noma, H., and Furukawa, T. A. (2018).
Prediction intervals for random-effects meta-analysis:
a confidence distribution approach.
\emph{Stat Methods Med Res}.
\emph{In press}.
\url{https://doi.org/10.1177/0962280218773520}.
}
\seealso{
\code{\link[=pima]{pima()}}
}
