% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cima-bc.r
\name{cima_bc}
\alias{cima_bc}
\title{Confidence interval based on Bartlett type corrections}
\usage{
cima_bc(y, se, alpha = 0.05)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{alpha}{the alpha level of the prediction interval}

\item{sigma}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{muhat}: the average treatment effect estimate \eqn{\hat{\mu}}.
\item \code{lci}, \code{lci}: the lower and upper confidence limits
                              \eqn{\hat{\mu}_l} and \eqn{\hat{\mu}_u}.
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
Returns a confidence interval for \eqn{\hat{\mu}} based on
Bartlett type corrections (Noma, 2011).
}
\examples{
data(sbp, package = "pimeta")
pimeta::cima_bc(sbp$y, sbp$sigmak)
}
\references{
Noma H. (2011)
Confidence intervals for a random-effects meta-analysis
based on Bartlett-type corrections.
\emph{Stat Med.}
\strong{30}(28): 3304-3312.
\url{https://doi.org/10.1002/sim.4350}
}
\seealso{
\code{\link[=cima]{cima}}.
}
