% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-mean.r
\name{convert_mean}
\alias{convert_mean}
\title{Converting means and standard deviations}
\usage{
convert_mean(n1, m1, s1, n2, m2, s2, pooled = FALSE)
}
\arguments{
\item{n1}{A vector of number of observations in experimental group}

\item{m1}{A vector of estimated mean in experimental group}

\item{s1}{A vector of standard deviation in experimental group}

\item{n2}{A vector of number of observations in experimental group}

\item{m2}{A vector of estimated mean in experimental group}

\item{s2}{A vector of standard deviation in experimental group}

\item{pooled}{logical; if \code{TRUE}, a pooled variance is used.
The default is \code{FALSE}.}
}
\value{
A \code{data.frame} of study data.
\itemize{
\item \code{y}: A numeric vector of the effect size estimates.
\item \code{se}: A numeric vector of the within studies standard error estimates.
}
}
\description{
Converting estimated means and standard deviations in
experimental and contorol groups to the effect size
estimates and the within studies standard errors vector
}
\examples{
require("flexmeta")
data("clbp")
dat <- convert_mean(clbp$n1, clbp$m1, clbp$s1, clbp$n2, clbp$m2, clbp$s2)
print(dat)
}
