% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-funnel.r
\name{funnel}
\alias{funnel}
\title{Funnel Plot}
\usage{
\method{plot}{pima}(
  x,
  title = "Funnel plot",
  base_size = 16,
  base_family = "",
  digits = 3,
  trans = c("identity", "exp")
)
}
\arguments{
\item{x}{`pima` or `cima` object to plot}

\item{title}{graph title}

\item{base_size}{base font size}

\item{base_family}{base font family}

\item{digits}{a value for digits specifies the minimum number
of significant digits to be printed in values.}

\item{trans}{transformation for logarithmic scale outcomes
(\code{"identity"} [default] or \code{"exp"}).}
}
\description{
Function for funnel plot of `pima` or `cima` objects.
}
\examples{
\donttest{
data(sbp, package = "pimeta")
piex <- pimeta::pima(sbp$y, sbp$sigmak, method = "HTS")
cairo_pdf("forestplot.pdf", width = 5, height = 5, family = "Arial")
funnel(piex, digits = 2, base_size = 10)
dev.off()
}
}
