% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pima-boot.r
\name{pima_boot}
\alias{pima_boot}
\alias{bootPI}
\title{A parametric bootstrap prediction interval}
\usage{
pima_boot(y, sigma, alpha = 0.05, B = 25000, maxit1 = 1e+05,
  eps = 10^(-10), lower = 0, upper = 1000, maxit2 = 1000,
  tol = .Machine$double.eps^0.25, rnd = NULL, parallel = FALSE,
  seed = NULL)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{sigma}{the within studies standard errors vector}

\item{alpha}{the alpha level of the prediction interval}

\item{B}{the number of bootstrap samples}

\item{maxit1}{the maximum number of iteration for the exact distribution function of \eqn{Q}}

\item{eps}{the desired level of accuracy for the exact distribution function of \eqn{Q}}

\item{lower}{the lower limit of random numbers of \eqn{\tau^2}}

\item{upper}{the upper limit of random numbers of \eqn{\tau^2}}

\item{maxit2}{the maximum number of iteration for numerical inversions}

\item{tol}{the desired level of accuracy for numerical inversions}

\item{rnd}{a vector of random numbers from the exact distribution of \eqn{\tau^2}}
}
\value{
\itemize{
\item \code{muhat}: the average treatment effect estimate \eqn{\hat{\mu}}.
\item \code{lci}, \code{lci}: the lower and upper confidence limits \eqn{\hat{\mu}_l} and \eqn{\hat{\mu}_u}.
\item \code{lpi}, \code{lpi}: the lower and upper prediction limits \eqn{\hat{c}_l} and \eqn{\hat{c}_u}.
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
A function for the parametric bootstrap PI
based on confidence distribution (Nagashima et al., 2018).
A parametric bootstrap confidence interval is also calculated
based on the same sampling method for bootstrap PI.
}
\examples{
data(sbp, package = "pimeta")
set.seed(20161102)
\donttest{pimeta::pima_boot(sbp$y, sbp$sigmak, B = 50000)}
}
\references{
Nagashima, K., Noma, H., and Furukawa, T. A. (2018).
Prediction intervals for random-effects meta-analysis:
a confidence distribution approach.
\emph{Stat Methods Med Res}.
\emph{In press}.
\url{https://doi.org/10.1177/0962280218773520}.

Hartung, J. (1999).
An alternative method for meta-analysis.
\emph{Biom J.}
\strong{41}(8): 901-916.
\url{https://doi.org/10.1002/(SICI)1521-4036(199912)41:8<901::AID-BIMJ901>3.0.CO;2-W}
}
\seealso{
\code{\link[=pima]{pima}}.
}
