% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{pimeta-package}
\alias{pimeta-package}
\title{Prediction Intervals for Random-Effects Meta-Analysis}
\author{
Kengo Nagashima, Hisashi Noma, and Toshi A. Furukawa
}
\keyword{package}
