% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cima-.r
\name{print.cima}
\alias{print.cima}
\title{Print Results}
\usage{
\method{print}{cima}(x, digits = 4, trans = c("identity", "exp"), ...)
}
\arguments{
\item{x}{print to display}

\item{digits}{a value for digits specifies the minimum number
of significant digits to be printed in values.}

\item{trans}{transformation for logarithmic scale outcomes
(\code{"identity"} [default] or \code{"exp"}).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} prints its argument and returns it invisibly (via \code{invisible(x)}).
}
