% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-.r
\name{print.pima_tau2h}
\alias{print.pima_tau2h}
\title{Print Results}
\usage{
\method{print}{pima_tau2h}(x, digits = 3, ...)
}
\arguments{
\item{x}{print to display}

\item{digits}{a value for digits specifies the minimum number
of significant digits to be printed in values.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} prints its argument and returns it invisibly (via \code{invisible(x)}).
}
