% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwchisq-.r
\name{pwchisq}
\alias{pwchisq}
\title{The Distribution of a Positive Linear Combination of Chiqaure Random Variables}
\usage{
pwchisq(x, lambda = 1, mult = 1, delta = 0, mode = 1,
  maxit1 = 1e+05, eps = 10^(-10))
}
\arguments{
\item{x}{numeric; value of x > 0 (\eqn{P[X \leq x]}).}

\item{lambda}{numeric vector; weights (\eqn{\lambda_1, \ldots, \lambda_K}).}

\item{mult}{integer vector; degrees of freedom (\eqn{m_1, \ldots, m_K}).}

\item{delta}{numeric vector; noncentraliy parameters (\eqn{\delta_1, \ldots, \delta_K}).}

\item{mode}{numeric; the mode of calculation (see Farabrother, 1984)}

\item{maxit1}{integer; the maximum number of iteration.}

\item{eps}{numeric; the desired level of accuracy.}
}
\value{
\itemize{
\item \code{prob}: the distribution function.
}
}
\description{
The distribution function for the distribution of a positive linear combination
of \eqn{\chi^2} random variables with the weights (\eqn{\lambda_1, \ldots, \lambda_K}),
degrees of freedom (\eqn{m_1, \ldots, m_K}), and noncentraliy parameters
(\eqn{\delta_1, \ldots, \delta_K}).
}
\examples{
# Farabrother (1984)'s Table 1, Q1 (The values were 0.0542, 0.4936, 0.8760)
pimeta::pwchisq( 1, lambda = c(6,3,1), mult = c(1,1,1), delta = c(0,0,0))
pimeta::pwchisq( 7, lambda = c(6,3,1), mult = c(1,1,1), delta = c(0,0,0))
pimeta::pwchisq(20, lambda = c(6,3,1), mult = c(1,1,1), delta = c(0,0,0))
}
\references{
Higgins, J. P. T., and Thompson, S. G. (2002).
Quantifying heterogeneity in a meta-analysis.
\emph{Stat Med.}
\strong{21}(11): 1539-1558.
\url{https://doi.org/10.1002/sim.1186}
}
