% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-areml.r
\name{tau2h_areml}
\alias{tau2h_areml}
\title{Approximate restricted maximum likelihood estimator for \eqn{\tau^2}}
\usage{
tau2h_areml(y, se, maxiter = 100)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}

\item{maxiter}{the maximum number of iterations}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
Returns an approximate restricted maximum likelihood estimator
for \eqn{\tau^2} (e.g., Thompson & Sharp, 1999).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_areml(sbp$y, sbp$sigmak)
}
\references{
Thompson, S. G., and Sharp, S. J. (1999).
Explaining heterogeneity in meta-analysis: a comparison of methods.
\emph{Stat Med.}
\strong{18}(20): 2693-2708.
}
