% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-bm.r
\name{tau2h_bm}
\alias{tau2h_bm}
\title{Bayes modal estimator for \eqn{\tau^2}}
\usage{
tau2h_bm(y, se)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
Returns the Chung--Rabe-Hesketh--Choi Bayes modal
estimator with a gamma prior \eqn{G(2, 10^{-4})}
for \eqn{\tau^2} (Chung et al., 2013).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_bm(sbp$y, sbp$sigmak)
}
\references{
Chung, Y. L., Rabe-Hesketh, S., and Choi, I-H. (2013).
Avoiding zero between-study variance estimates
in random-effects meta-analysis.
\emph{Stat Med.}
\strong{32}(23): 4071-4089.
}
