% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-dl.r
\name{tau2h_dl}
\alias{tau2h_dl}
\title{DerSimonian--Laird estimator for \eqn{\tau^2}}
\usage{
tau2h_dl(y, se)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
Returns the DerSimonian--Laird estimator
for \eqn{\tau^2} (DerSimonian & Laird, 1986).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_dl(sbp$y, sbp$sigmak)
}
\references{
DerSimonian, R., and Laird, N. (1986).
Meta-analysis in clinical trials.
\emph{Control Clin Trials.}
\strong{7}(3): 177-188.
}
