% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-eb.r
\name{tau2h_eb}
\alias{tau2h_eb}
\title{Empirical Bayes estimator for \eqn{\tau^2}}
\usage{
tau2h_eb(y, se, maxiter = 100)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}

\item{maxiter}{the maximum number of iterations}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
Returns an Empirical Bayes estimator
for \eqn{\tau^2} (Morris, 1983).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_eb(sbp$y, sbp$sigmak)
}
\references{
Morris, C. N. (1983).
Parametric empirical Bayes inference: theory and 
applications.
\emph{J Am Stat Assoc.}
\strong{78}(381): 47-55.
}
