% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-hm.r
\name{tau2h_hm}
\alias{tau2h_hm}
\title{Hartung--Makambi estimator for \eqn{\tau^2}}
\usage{
tau2h_hm(y, se)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
A subroutine of the Hartung--Makambi estimator
for \eqn{\tau^2} (Hartung & Makambi, 1983).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_hm(sbp$y, sbp$sigmak)
}
\references{
Hartung, J., and Makambi, K. H. (2003).
Reducing the number of unjustified significant results
in meta-analysis. 
\emph{Commun Stat Simul Comput.}
\strong{32}(4): 1179-1190.
}
