% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-hs.r
\name{tau2h_hs}
\alias{tau2h_hs}
\title{Hunter--Schmidt estimator for \eqn{\tau^2}}
\usage{
tau2h_hs(y, se)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
Returns the Hunter--Schmidt estimator
for \eqn{\tau^2} (Hunter & Schmidt, 2004).
This estimator has negative bias (Viechtbauer, 2005).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_hs(sbp$y, sbp$sigmak)
}
\references{
Hunter, J. E., and Schmidt, F. L. (2004).
\emph{Methods of Meta-Analysis: Correcting Error and Bias in Research Findings. 2nd edition.}
Sage Publications, Inc.

Viechtbauer, W. (2005).
Bias and efficiency of meta-analytic variance
estimators in the random-effects Model.
\emph{J Educ Behav Stat.}
\strong{30}(3): 261-293.
}
