% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-pm.r
\name{tau2h_pm}
\alias{tau2h_pm}
\title{Paule--Mandel estimator for \eqn{\tau^2}}
\usage{
tau2h_pm(y, se)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
Returns the Paule--Mandel estimator
for \eqn{\tau^2} (Paule & Mandel, 1982).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_pm(sbp$y, sbp$sigmak)
}
\references{
Paule, R. C., and Mandel, K. H. (1982).
Consensus values and weighting factors.
\emph{J Res Natl Inst Stand Techno.}
\strong{87}(5): 377-385.
}
