% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-reml.r
\name{tau2h_reml}
\alias{tau2h_reml}
\title{Restricted maximum likelihood estimator for \eqn{\tau^2}}
\usage{
tau2h_reml(y, se, maxiter = 100)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}

\item{maxiter}{the maximum number of iterations}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
A subroutine of a restricted maximum likelihood estimator
for \eqn{\tau^2} (e.g., DerSimonian & Laird, 1986).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_reml(sbp$y, sbp$sigmak)
}
\references{
DerSimonian, R., and Laird, N. (1986).
Meta-analysis in clinical trials.
\emph{Control Clin Trials.}
\strong{7}(3): 177-188.
}
