% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-sj.r
\name{tau2h_sj}
\alias{tau2h_sj}
\title{Sidik--Jonkman estimator for \eqn{\tau^2}}
\usage{
tau2h_sj(y, se)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
A subroutine of the Sidik--Jonkman estimator
for \eqn{\tau^2} (Sidik & Jonkman, 2005).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_sj(sbp$y, sbp$sigmak)
}
\references{
Sidik, K., and Jonkman, J. N. (2005).
Simple heterogeneity variance estimation for meta-analysis.
\emph{J R Stat Soc Ser C Appl Stat.}
\strong{54}(2): 367-384.
}
