% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-sj2.r
\name{tau2h_sj2}
\alias{tau2h_sj2}
\title{Sidik--Jonkman improved estimator for \eqn{\tau^2}}
\usage{
tau2h_sj2(y, se)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
A subroutine of the Sidik--Jonkman improved estimator
for \eqn{\tau^2} (Sidik & Jonkman, 2007).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_sj2(sbp$y, sbp$sigmak)
}
\references{
Sidik, K., and Jonkman, J. N. (2007).
A comparison of heterogeneity variance estimators
in combining results of studies.
\emph{Stat Med.}
\strong{26}(9): 1964-1981.
}
