% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau2h-vc.r
\name{tau2h_vc}
\alias{tau2h_vc}
\title{Variance component type estimator for \eqn{\tau^2}}
\usage{
tau2h_vc(y, se)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{se}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
Returns the variance component type estimator
for \eqn{\tau^2} (Hedges, 1983).
}
\examples{
data(sbp, package = "pimeta")
pimeta::tau2h_vc(sbp$y, sbp$sigmak)
}
\references{
Hedges, L. V. (1983).
A random effects model for effect sizes.
\emph{Psychol Bull.}
\strong{93}(2): 388-395.
}
