% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cima-pl.r
\name{cima_pl}
\alias{cima_pl}
\alias{htsdl}
\title{Brockwell--Gordon confidence interval}
\usage{
cima_pl(y, se, alpha = 0.05)
}
\arguments{
\item{y}{the effect size estimates vector}

\item{alpha}{the alpha level of the prediction interval}

\item{sigma}{the within studies standard errors vector}
}
\value{
\itemize{
\item \code{muhat}: the average treatment effect estimate \eqn{\hat{\mu}}.
\item \code{lci}, \code{lci}: the lower and upper confidence limits
                              \eqn{\hat{\mu}_l} and \eqn{\hat{\mu}_u}.
\item \code{tau2h}: the estimate for \eqn{\tau^2}.
}
}
\description{
Returns the Brockwell--Gordon confidence interval
for \eqn{\hat{\mu}} (Brockwell & Gordon, 2007).
}
\examples{
data(sbp, package = "pimeta")
pimeta::cima_pl(sbp$y, sbp$sigmak)
}
\references{
Hardy, R. J., and Thompson, S. G. (1996).
A likelihood approach to meta-analysis with random effects.
\emph{Stat Med.}
\strong{15}(6): 619-629.
\url{https://doi.org/10.1002/(SICI)1097-0258(19960330)15:6<619::AID-SIM188>3.0.CO;2-A}.
}
\seealso{
\code{\link[=cima]{cima}}.
}
