% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{pimeta-package}
\alias{pimeta-package}
\title{Prediction Intervals for Random-Effects Meta-Analysis}
\description{
Prediction Intervals for Random-Effects Meta-Analysis.
}
\author{
Author: Kengo Nagashima, Hisashi Noma, and Toshi A. Furukawa

Maintainer: Kengo Nagashima \email{nshi@chiba-u.jp}
}
\keyword{package}
