% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cima-.r
\name{plot.cima}
\alias{plot.cima}
\title{Plot Results}
\usage{
\method{plot}{cima}(x, y = NULL, title = "Forest plot",
  base_size = 16, base_family = "", ...)
}
\arguments{
\item{x}{`cima` object to plot}

\item{y}{is not used}

\item{title}{graph title}

\item{base_size}{base font size}

\item{base_family}{base font family}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A function for plotting of `cima` objects.
}
