% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pima-.r
\name{plot.pima}
\alias{plot.pima}
\title{Plot Results}
\usage{
\method{plot}{pima}(x, y = NULL, title = "Forest plot",
  base_size = 16, base_family = "", digits = 3, studylabel = NULL,
  ntick = NULL, trans = c("identity", "exp"), ...)
}
\arguments{
\item{x}{`pima` object to plot}

\item{y}{is not used}

\item{title}{graph title}

\item{base_size}{base font size}

\item{base_family}{base font family}

\item{digits}{a value for digits specifies the minimum number
of significant digits to be printed in values.}

\item{studylabel}{labels for each study}

\item{ntick}{the number of x-axis ticks}

\item{trans}{transformation for logarithmic scale outcomes
(\code{"identity"} [default] or \code{"exp"}).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A function for plotting of `pima` objects.
}
\examples{
data(sbp, package = "pimeta")
piex <- pimeta::pima(sbp$y, sbp$sigmak, method = "HTS")
cairo_pdf("forestplot.pdf", width = 6, height = 3, family = "Arial")
plot(piex, digits = 2, base_size = 10, studylabel = sbp$label)
dev.off()
}
